+++ Models/TipoIVA.cs
@@ -4,7 +4,7 @@ public class TipoIVA{
 
     public int? idTipoIVA {get; set;}
     public string? descripcion {get; set;}
-    public double? tasa {get; set;}
+    public decimal? tasa {get; set;}
     public DateTime? fechaCreacion {get; set;}
     public bool activo {get; set;}
     
@@ -14,7 +14,7 @@ public class TipoIVA{
         activo = true;
     }
 
-    public TipoIVA(int _idTipoIVA, string _descripcion, double _tasa, DateTime _fechaCreacion)
+    public TipoIVA(int _idTipoIVA, string _descripcion, decimal _tasa, DateTime _fechaCreacion)
     {
         idTipoIVA = _idTipoIVA;
         descripcion = _descripcion;
